registration final
