+++ src/api/models.py
@@ -23,10 +23,8 @@ class UserRegister(db.Model):
 # class User(db.Model):
 #     id = db.Column(db.Integer, primary_key=True)
 #     user_register_id = db.Column(db.Integer, db.ForeignKey("user_register.id"), nullable=False)
-
 #     def __repr__(self):
 #         return f'<User {self.id}>'
-
 #     def serialize(self):
 #         return {
 #             "id": self.id,
