+++ src/api/routes.py
@@ -6,6 +6,12 @@ from api.models import db, UserRegister, Restaurant
 from api.utils import generate_sitemap, APIException
 from flask_cors import CORS
 
+
+import hashlib
+import random
+from flask_jwt_extended import create_access_token
+
+
 from flask_jwt_extended import get_jwt_identity
 from flask_jwt_extended import jwt_required
 
@@ -76,4 +82,59 @@ def create_restaurant():
     return jsonify({"message": "Please Register to save the favourite restaurant!"}), 400
 
 
+@api.route('/register', methods=['POST'])
+def register_user():
+    data = request.json
+    username = data.get("username")
+    email = data.get("email")
+    password = data.get("password")
 
+    # Check if username, email, and password are provided
+    if not username or not email or not password:
+        return jsonify({"message": "Username, email, and password are required"}), 400
+
+    # Check if user already exists
+    if UserRegister.query.filter_by(email=email).first():
+        return jsonify({"message": "User with this email already exists"}), 400
+
+    # Create new user
+    hashed_password = hashlib.sha256(password.encode()).hexdigest()
+    user_register = UserRegister(user_name=username, email=email, password=hashed_password)
+    db.session.add(user_register)
+    db.session.commit()
+
+    return jsonify({"message": "User registered successfully"}), 201
+
+# User creation endpoint
+# @app.route('/create_user', methods=['POST'])
+# def create_user():
+#     data = request.json
+#     user_register_id = data.get("user_register_id")
+#     username = data.get("username")
+
+#     # Create new user
+#     user = User(user_register_id=user_register_id, user_name=username)
+#     db.session.add(user)
+#     db.session.commit()
+
+    # return jsonify({"message": "User created successfully"}), 201
+
+# Main execution block
+
+@api.route("/login", methods=["POST"])
+def login():
+    data = request.json
+    email = data.get("email")
+    password = data.get("password")
+
+    # Check if username, email, and password are provided
+    if not email or not password:
+        return jsonify({"message": "Username, email, and password are required"}), 400
+    hashed_password = hashlib.sha256(password.encode()).hexdigest()
+
+    # Check if user already exists
+    if UserRegister.query.filter_by(email=email, password = hashed_password).first():
+        token = create_access_token(identity=email)
+        return jsonify({"message": "Welcome to Your favourite restaurant lists", "token":token}), 200
+
+    return jsonify({"message": "Please provide proper credentials."}), 400
